#3 rename test_cases_all.py to test_cases.py
